Rename id to folderId for clarity
