+++ frontend/pages/workspaces/[workspaceId]/folders/folders.graphql
